Game water-level timer, layout/safe-area fixes, Lottie renderer getFloat
Games:
- Replace the small vertical glass with a full-screen "water level" timer
  behind the gameplay (native:stack back layer): a tinted fill that drains a
  step each second, with a bright waterline and a gentle looping slosh on the
  body + surface so it isn't static. Applies to both games.
- Fix the game layout: the content column now has a real height inside the
  stack, so the prompt no longer collapses behind the options.
- Add safe-area to the immersive game screens so the close button clears the
  notch.
- Remove the hero illustration from the game detail (instruction) header.

Lottie plugin:
- Swift/Kotlin renderers read node props via getFloat (GenericProps has no
  getDouble); iOS converts to CGFloat for lottie's currentProgress/speed.
  (The iOS build now clears package resolution + CocoaPods; on-device render
  still to be confirmed.)

Suite green (125 tests); Pint clean.

+++ plugins/nativephp-lottie/resources/android/LottieRenderer.kt
@@ -25,8 +25,8 @@ object LottieRenderer {
         }
 
         val loop = p.getBool("loop")
-        val speed = p.getDouble("speed", 1.0).toFloat()
-        val progress = p.getDouble("progress", -1.0).toFloat()
+        val speed = p.getFloat("speed", 1.0f)
+        val progress = p.getFloat("progress", -1.0f)
 
         val composition by rememberLottieComposition(
             LottieCompositionSpec.Asset("animations/$source.json")
